Update to split scripts according to ansible roles.
Which seems to be the preferred way to split things instead of tasks.

There is still a problem of not being able to specify the roles unless
you change the hard-coded value in the Vagrant file, but that's not a
huge priority right now.

One thing I would still like to do is further split the settings by
type, but again not a huge priority.

Also:
- Added checks for local settings files before copying.
- Added guards to the lein playbook.
- Added postgresql.